data_acquisition/requirements.txt
alpaca-trade-api
ib_insync
python-binance
pandas

